requirements.txt
beautifulsoup4==4.12.3
folium==0.19.2
python-telegram-bot==21.9
Requests==2.32.3
selenium==4.27.1
lxml==5.3.0